Rename source directory
Redundant names create overly long names.  Rename directories in order
to keep overall path names more reasonable in length and avoid
redundencies in the pathname.  Update the tests to accommodate these
changes.
